Refactored the code base to encapsulate functionality into separate class files
